+++ CHANGELOG.md
@@ -23,6 +23,12 @@ Here we write upgrading notes and make them as straightforward as possible.
 - A short description for fixed item n
 
 
+## [v0.5.1] - 2025-06-30
+
+### Added
+- `scan()` for partially matching keys
+- `free()` for freeing the scanned keys
+
 ## [v0.5.0] - 2025-06-22
 
 Initial bare minimum implementation (Sync only).